refactor(auto_source): fold Anchorless and deepen DomClustering
- Fold transparent Discovery::Anchorless passthrough into callers
- Consolidate class-based and structure-based DOM candidate clustering into deep Discovery::DomClustering
- Thread identity cache through ListCandidates to avoid redundant ancestor traversals
- Update CONTEXT.md and replace separate specs with dom_clustering_spec
